Initial work on moving media controller to the backend
