+++ modules/backend/controllers/media/index.htm
@@ -1,0 +1,7 @@
+<?= Block::put('head') ?><?= Block::endPut() ?>
+
+<?= Block::put('body') ?>
+    <?= Form::open(['class'=>'layout', 'onsubmit'=>'return false']) ?>
+        <?= $this->widget->manager->render() ?>
+    <?= Form::close() ?>
+<?= Block::endPut() ?>
